add theme light

+++ lib/config/constants.dart
@@ -2,5 +2,5 @@ class Constans {
   static const String klinkPro = '';
   static const double deviceWidth = 375;
   static const double deviceHeight = 812;
-  static const double splashDuration = 3;
+  static const int splashDuration = 3;
 }
